refactor(web): merge neighbour table modals into one TableModal
CreateTableModal and EditTableModal shared near-identical modal chrome.
Merge them into a single mode-driven TableModal, preserving each mode's
exact fields and behaviour.
