+++ web/src/pages/operators/neighbours/types.ts
@@ -1,5 +1,3 @@
-import type { NeighbourTableInfo } from '../../../api/neighbours';
-
 export type SortableColumn =
     | 'next_hop'
     | 'link_addr'
@@ -20,17 +18,3 @@ export interface SortState {
 export const DEFAULT_SORT: SortState = { column: 'state', direction: 'asc' };
 
 export const MERGED_TAB = '__merged__';
-
-export interface CreateTableModalProps {
-    open: boolean;
-    onClose: () => void;
-    onCreate: (name: string, defaultPriority: number) => Promise<void>;
-    existingNames: string[];
-}
-
-export interface EditTableModalProps {
-    open: boolean;
-    onClose: () => void;
-    onSave: (name: string, defaultPriority: number) => Promise<void>;
-    tableInfo: NeighbourTableInfo | null;
-}
